Fixed a bug: wrong update of arrays from SSH-aerosol.
For low liquid water content in the street segments, munich calls ssh-aerosol and get the concentration from ssh-aerosol.
When the liquid water content is higher than a given threshold, the program calls an aqueous model (simple or vsrm) instead of ssh-aerosol.
Even though the program doesn't call ssh-aerosol, the concentration was updated from ssh-aerosol which may correspond to a concentration of a previous time step.